Update 0.2 - Inserting Songs
Continued to recycle code from original insert.php to implement insert_song.php.  Not longer need insert.php file but left it until later to delete after project is done just in case.

Implemented a way to insert songs into the database, which will also add artists and albums if not already in the database.

One note to keep in mind for the future that worried me is deleting entries (what happens if you delete an artist who is linked to a bunch of albums and songs? how do you relink that?) One possible solution is that any time someone wants to delete something, instead they aren't truely deleting it but rather replacing it with another name. Or instead make deleting impossible, and replacing the only way.